Restructured docs directories
