+++ src/components/DocsView/pages/resources/ResourcesPage.tsx
