+++ src/components/PricingView/PricingViewTable.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import Header from '../Header'
+import Header from '../DocsView/components/Header/Header'
 import Footer from '../Footer/Footer'
 import { $p, $v } from 'graphcool-styles'
 import styled from 'styled-components'
