+++ src/main.tsx
@@ -8,11 +8,11 @@ import AboutView from './components/AboutView/AboutView'
 import DocsView from './components/DocsView/DocsView'
 import FAQ from './components/DocsView/FAQ/FAQ'
 import * as FastClick from 'fastclick'
-import QuickstartPage from './pages/quickstart/QuickstartPage'
-import ReferencePage from './pages/reference/ReferencePage'
-import ResourcesPage from './pages/resources/ResourcesPage'
-import BlogPage from './pages/blog/BlogPage'
-import CommunityPage from './pages/community/CommunityPage'
+import QuickstartPage from './components/DocsView/pages/quickstart/QuickstartPage'
+import ReferencePage from './components/DocsView/pages/reference/ReferencePage'
+import ResourcesPage from './components/DocsView/pages/resources/ResourcesPage'
+import BlogPage from './components/DocsView/pages/blog/BlogPage'
+import CommunityPage from './components/DocsView/pages/community/CommunityPage'
 import ContentHandler from './components/DocsView/ContentHandler/ContentHandler'
 import ApolloClient, {createNetworkInterface} from 'apollo-client'
 import { ApolloProvider } from 'react-apollo'
